examples/gradio/requirements.txt
openai==1.6.1
python-dotenv==1.0.0
gradio==4.13.0
